+++ src/decorators/has-many.decorator.ts
@@ -1,12 +1,15 @@
+import { HasManyRelationshipMetadata } from '../constants/symbols';
+
 export function HasMany(config: any = {}) {
   return function (target: any, propertyName: string | symbol) {
-    const annotations = Reflect.getMetadata('HasMany', target) || [];
+    console.log(HasManyRelationshipMetadata);
+    const annotations = Reflect.getMetadata(HasManyRelationshipMetadata, target) || [];
 
     annotations.push({
       propertyName,
       relationship: config.key || propertyName
     });
 
-    Reflect.defineMetadata('HasMany', annotations, target);
+    Reflect.defineMetadata(HasManyRelationshipMetadata, annotations, target);
   };
 }